webapp/requirements.txt
Django==1.10.4
django-appconf==1.0.2
django-bootstrap-pagination==1.6.2
django-haystack==2.5.1
elasticsearch==5.0.1
mysqlclient==1.3.9
urllib3==1.19.1
